Funnel all changes through -modifyWithBlock:
The old way funneled through -setValue:forKeyPath:. If there is no
document, -modifyWithBlock: now applies the change immediately; if there
is a document, the modification block is made root-relative and passed
up.

We don't call -commitModificationsKeepingChanges: on collections we
didn't send -beginModifications to (perhaps because the collection was
added as the modification).